Run Chrome and Firefox tests concurrently
This involved passing in a unique tag to be used in generating instance
names. In the GitLab CI environment this will include the browser brand.

Also skipping operating system updates to launch instances faster.

For reliability purposes we no longer save and load the local storage.
Instead we have a single step which adds a Jetstream cloud account in
the beginning of the 'launch' and 'delete' scenarios.